+++ RELEASE-NOTES.md
@@ -20,6 +20,7 @@
 ## Deprecations
 
 ## New additions
+* Added `!spool` command to SQL REPL for writing query output to a file (`!spool <filename>` to start, `!spool off` to stop).
 
 ## Fixes and improvements
 * all authenticators (including `snowflake-jwt`, `username_password_mfa`, `workload_identity`) are now case-insensitive.